Minor project tweaks
