+++ .gitignore
@@ -1,4 +1,2 @@
-
 .DS_Store
-
-star-map.xcodeproj/project.xcworkspace/xcuserdata/yelienakhaletska.xcuserdatad/
+star-map.xcodeproj/project.xcworkspace/xcuserdata/
