+++ star-map/SupportingFiles/Assets.xcassets/AppIcon.appiconset/Contents.json
