+++ star-map/SupportingFiles/Assets.xcassets/Contents.json
